Refactor how Positions are handled
This change introduces structs:
ScreenPosition: A position on the screen
WorldPosition: A position in a viewport (e.g. a camera position)
MapPosition: A tile (X, Y) on a map
Position: A tile (X, Y, Z) on a map

+++ vulkan-tutorial/map.cpp
@@ -190,7 +190,7 @@ MapIterator &MapIterator::operator++()
     tileIndex = 0;
     floorIndex = 0;
 
-    uint32_t size = current.node->nodes.size();
+    size_t size = current.node->nodes.size();
     // This node is finished
     if (current.cursor == size)
     {
